tests/requirements.txt
langchain>=0.2.3,<0.3
langchain-openai>=0.1.7,<0.2
python-dotenv
pytest